.github/workspace/app/requirements.txt
absl-py==2.0.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
audioread==3.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
exceptiongroup==1.1.3
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.43.1
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
joblib==1.1.0
JPype1 @ file:///C:/kd/setup/JPype1-1.4.0-cp39-cp39-win_amd64.whl#sha256=3dcfee215e949acd79eccffbfae62b88c688200db3b326eef1cdca58bb08de91
jupyter_client==8.5.0
jupyter_core==5.5.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.9.1
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.4.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
msgpack==1.0.7
mysqlclient==2.2.0
nest-asyncio==1.5.8
numba==0.58.0
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==1.4.1
parso==0.8.3
Pillow==10.1.0
platformdirs==3.11.0
pooch==1.8.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikeras==0.12.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.13.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.4
stack-data==0.6.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.2.0
torch==1.9.0
torchvision==0.10.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Wave==0.0.2
wcwidth==0.2.9
webcolors==1.13
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
